refactor: refactor securiy page, add security form hook
